+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## v1.0.0 - 2025-05-04 - Long overdue 1.0
 
 * Address pending octoDNS 2.x deprecations, require minimum of 1.5.x
+* Add hcloud backend with Strategy Pattern for Cloud Zones API support
+* Fix hcloud zone creation race condition by caching zone objects
 
 ## v0.0.3 - 2023-02-08 - AKA